+++ composer.json
@@ -7,10 +7,12 @@
     }
   ],
   "require": {
-    "php": ">=5.5.0",
-    "phpunit/phpunit": "5.0.8"
+    "php": ">=5.5.0"
   },
-    "autoload": {
+  "require-dev": {
+    "phpunit/phpunit": "5.0.*"
+  },
+  "autoload": {
     "psr-4": {
       "Calculator\\": "src/Calculator"
     }
