+++ src/Calculator/Calculator.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: dmitry
+ * Date: 03.11.15
+ * Time: 22:41
+ */
+namespace Calculator;
+
+class Calculator
+{
+    //  public $a = 2;
+    // public $b = 3;
+
+    public function add($a, $b)
+    {
+        return $a + $b;
+    }
+
+    public function multipl($a, $b)
+    {
+        return $a * $b;
+    }
+}
